Fixing hydration for carousel
