+++ components/blocks/carousel.tsx
@@ -2,22 +2,32 @@
 import Image from "next/image";
 import { useRouter } from "next/navigation";
 import * as React from "react";
-import { isMobile } from "react-device-detect";
 import { tinaField } from "tinacms/dist/react";
 
-import type { Template } from "tinacms";
-
+import { isMobile } from "react-device-detect";
 import { Container } from "../util/container";
 import { Section } from "../util/section";
 
 import "react-responsive-carousel/lib/styles/carousel.min.css";
 
 import { Carousel as CarouselImplementation } from "react-responsive-carousel";
+import { carouselBlock } from "./carousel.schema";
 
 export const Carousel = ({ data }) => {
   const router = useRouter();
 
-  if (!data.showOnMobileDevices && isMobile) {
+  // Handle the mobile check after hydration
+  const [shouldRender, setShouldRender] = React.useState(true);
+
+  React.useEffect(() => {
+    if (!data.showOnMobileDevices && isMobile) {
+      setShouldRender(false);
+    } else {
+      setShouldRender(true);
+    }
+  }, [data.showOnMobileDevices]);
+
+  if (!shouldRender) {
     return null;
   }
 
@@ -132,94 +142,4 @@ const createCarouselIndicator = (onClickHandler, isSelected, index, label) => {
       aria-label={`${label} ${index + 1}`}
     />
   );
-};
-
-export const carouselBlock = {
-  items: {
-    value: "items",
-    label: "label",
-    link: "link",
-    imgSrc: "imgSrc",
-  },
-  delay: "delay",
-};
-
-export const carouselBlockSchema: Template = {
-  name: "Carousel",
-  label: "Carousel",
-  ui: {
-    previewSrc: "/images/thumbs/tina/carousel.jpg",
-  },
-  fields: [
-    {
-      label: "Items",
-      name: carouselBlock.items.value,
-      type: "object",
-      list: true,
-      ui: {
-        defaultItem: {
-          label: "Item description",
-          link: "/",
-          openIn: "sameWindow",
-        },
-        itemProps: (item) => ({ label: item.label }),
-      },
-      fields: [
-        {
-          type: "string",
-          label: "Label",
-          name: carouselBlock.items.label,
-        },
-        {
-          type: "string",
-          label: "URL",
-          name: carouselBlock.items.link,
-          description:
-            "If link contains ssw.com.au, you can skip the full URL and just use the path. e.g. /services",
-        },
-        {
-          type: "string",
-          label: "Open in",
-          name: "openIn",
-          description:
-            "If it is external link, please select 'New window' option.",
-          options: [
-            { label: "Same window", value: "sameWindow" },
-            { label: "Modal", value: "modal" },
-            { label: "New window", value: "newWindow" },
-          ],
-        },
-        {
-          type: "image",
-          label: "Image",
-          name: carouselBlock.items.imgSrc,
-          // eslint-disable-next-line @typescript-eslint/ban-ts-comment
-          // @ts-ignore
-          uploadDir: () => "carousel",
-        },
-      ],
-    },
-    {
-      type: "string",
-      label: "Background Color",
-      name: "backgroundColor",
-      options: [
-        { label: "Default", value: "default" },
-        { label: "Light Gray", value: "lightgray" },
-        { label: "Red", value: "red" },
-        { label: "Black", value: "black" },
-      ],
-    },
-    {
-      type: "number",
-      label: "Delay (Seconds)",
-      name: carouselBlock.delay,
-      required: true,
-    },
-    {
-      type: "boolean",
-      label: "Show on mobile devices",
-      name: "showOnMobileDevices",
-    },
-  ],
 };
