+++ tina/collections/company.tsx
@@ -1,6 +1,6 @@
 import * as Schemas from "../../components/blocks";
 
-import { carouselBlockSchema } from "../../components/blocks/carousel";
+import { carouselBlockSchema } from "../../components/blocks/carousel.schema";
 import { fixedColumnsSchema } from "../../components/blocks/fixedColumns";
 import { sectionHeaderSchema } from "../../components/blocks/sectionHeader";
 import { tableBlockSchema } from "../../components/blocks/tableLayout";
